management/server/requirements.txt
Flask==3.1.0
flask_cors==5.0.1
mysql-connector-python==9.2.0
pycryptodomex==3.20.0
tabulate==0.9.0
Werkzeug==3.1.3
PyJWT==2.10.1
dotenv==0.9.9
magic-pdf[full]==1.3.0
transformers==4.49.0
elasticsearch==8.12.0
minio==7.2.4
strenum==0.4.15
peewee==3.17.1
logto==0.2.1
